Extract OIM loading into core Arelle module
