+++ pyproject.toml
@@ -294,6 +294,7 @@ module = [
     'arelle.formula.FormulaConsisAsser',
     'arelle.formula.FormulaEvaluator',
     'arelle.formula.ValidateFormula',
+    'arelle.oim.Load',
     'arelle.plugin.CacheBuilder',
     'arelle.plugin.EdgarRendererAllReports',
     'arelle.plugin.SECCorrespondenceLoader',
@@ -358,7 +359,6 @@ module = [
     'tests.plugin.testcaseIxExpectedHtmlFixup',
     'tests.unit_tests.arelle.conftest',
     'tests.unit_tests.arelle.formula.test_fact_aspects_cache',
-    'tests.unit_tests.arelle.plugin.test_loadfromoim',
     'tests.unit_tests.arelle.plugin.test_plugin_imports',
     'tests.unit_tests.arelle.test_betafeatures',
     'tests.unit_tests.arelle.test_cntlr',
